+++ packages/ui/src/components/widgets/OnrampWidget/widget/index.tsx
@@ -66,6 +66,7 @@ const OnrampWidget: FC<OnrampWidgetProps> = ({
   supportedWalletVMs,
   moonPayThemeId,
   moonPayThemeMode,
+  defaultToken,
   onConnectWallet,
   onLinkNewWallet,
   onSetPrimaryWallet,
@@ -83,6 +84,7 @@ const OnrampWidget: FC<OnrampWidgetProps> = ({
       linkedWallets={linkedWallets}
       multiWalletSupportEnabled={multiWalletSupportEnabled}
       moonPayApiKey={moonPayApiKey}
+      defaultToken={defaultToken}
     >
       {({
         displayCurrency,
@@ -182,19 +184,35 @@ const OnrampWidget: FC<OnrampWidgetProps> = ({
                   placeholder={`   0`}
                   onChange={(e) => {
                     const input = e.target as any
+
                     setInputValue(input.value)
                     if (displayCurrency) {
                       setTimeout(() => {
-                        const cursorPosition = input.selectionStart
-                        if (cursorPosition >= input.value.length) {
-                          input.setSelectionRange(0, 0)
-                        }
+                        const numericValue = input.value.match(/[\d.]+/g)
+                        const numericValueLength =
+                          numericValue !== null
+                            ? numericValue.join('').length
+                            : 0
+                        input.setSelectionRange(
+                          numericValueLength,
+                          numericValueLength
+                        )
                       }, 0)
                     }
                   }}
                   onKeyDown={(e) => {
                     const input = e.target as HTMLInputElement
                     const cursorPosition = input.selectionStart
+
+                    // Prevent multiple decimals
+                    if (
+                      e.key === '.' &&
+                      (input.value.match(/\./g) || []).length > 0
+                    ) {
+                      e.preventDefault() // Prevent the key press if there's already a decimal
+                      return
+                    }
+
                     if (e.key === 'ArrowLeft' && cursorPosition !== null) {
                       const valueBeforeCursor = input.value.substring(
                         0,